requirements.dev.txt
-r requirements.txt
djlint==1.34.1
black==24.4.2
isort==5.13.2
pylint==3.3.1
django-debug-toolbar==4.4.6